fotografias fix

+++ app/Http/Controllers/FotografiaController.php
@@ -90,7 +90,7 @@ class FotografiaController extends Controller
     {
         $validacio = Validator::make($request->all(), [
             'ID_FOTO' => 'required|integer',
-            'FOTO' => 'required|string',
+            'FOTO' => 'required',
             'FK_ID_ALLOTJAMENT' => 'required|integer'
         ]);
 
@@ -102,8 +102,7 @@ class FotografiaController extends Controller
             $foto->FOTO = $request->FOTO;
             $foto->FK_ID_ALLOTJAMENT = $request->FK_ID_ALLOTJAMENT;
             $foto->save();
-            return response()->json(['success' => 'Fotografia creada correctament',
-                'result' => $foto], 201);
+            return response()->json($foto, 201);
         }
     }
 
@@ -190,14 +189,19 @@ class FotografiaController extends Controller
      */
     public function updateFotografia(Request $request, $id)
     {
-        $validacio = Validator::make($request->all(), [
+        $regles = [
             'ID_FOTO' => 'required|integer',
-            'FOTO' => 'required|string',
+            'FOTO' => 'required',
             'FK_ID_ALLOTJAMENT' => 'required|integer'
-        ]);
+        ];
 
-        if ($validacio->fails()) {
-            return response()->json(['error' => $validacio->errors()->first()], 400);
+        $missatge = [
+            'required' => 'El camp :attribute és obligatori',
+            'integer' => 'El camp :attribute ha de ser un número enter'
+        ];
+        $foto = Validator::make($request->all(), $regles, $missatge);
+        if($foto->fails()){
+            return response()->json(['error' => $foto->errors()->first()], 400);
         } else {
             try {
                 $foto = Fotografia::findOrFail($id);
@@ -205,12 +209,13 @@ class FotografiaController extends Controller
                 $foto->FOTO = $request->FOTO;
                 $foto->FK_ID_ALLOTJAMENT = $request->FK_ID_ALLOTJAMENT;
                 $foto->save();
-                return response()->json(['success' => 'Fotografia modificada correctament',
+                return response()->json(['success' => 'Fotografia actualitzada correctament',
                     'result' => $foto], 200);
             } catch (ModelNotFoundException $e) {
                 return response()->json(['error' => 'No s\'ha trobat la fotografia'], 404);
             }
         }
+
     }
 
     /**
@@ -246,11 +251,11 @@ class FotografiaController extends Controller
      */
     public function deleteFotografia($id)
     {
-        try {
-            $foto = Fotografia::findOrFail($id);
+        $foto = Fotografia::find($id);
+        if ($foto) {
             $foto->delete();
             return response()->json(['success' => 'Fotografia eliminada correctament'], 200);
-        } catch (ModelNotFoundException $e) {
+        } else {
             return response()->json(['error' => 'No s\'ha trobat la fotografia'], 404);
         }
     }
